In the C++ version I modified the varibale types to imporve speed due to not having to do type conversions on runtime

+++ 44-Pentagon_Numbers/main.cpp
@@ -2,12 +2,12 @@
 #include <cmath>
 #include <vector>
 
-bool isPentagonal(long long x) {
+bool isPentagonal(long x) {
     double n = (1 + sqrt(1 + 24 * x)) / 6;
-    return n == (long long)n;
+    return n == (long) n;
 }
 
-long long pentagonal(int n) {
+long pentagonal(int n) {
     return n * (3LL * n - 1) / 2;
 }
 
@@ -16,12 +16,12 @@ int main() {
     int n = 1;
 
     while (true) {
-        long long P_n = pentagonal(n);
+        long P_n = pentagonal(n);
         pentagonals.push_back(P_n);
 
         for (auto P_j : pentagonals) {
             if (P_j == P_n) continue;
-            long long P_k = P_n;
+            long P_k = P_n;
             if (isPentagonal(P_k - P_j) && isPentagonal(P_k + P_j)) {
                 std::cout << "D = " << P_k - P_j << std::endl;
                 return 0;
